+++ CHANGELOG.md
@@ -54,12 +54,12 @@ See [Issue Backlog](../../issues) and [Roadmap](../../milestones).
   - Support 4 Quick Test options, based on skills in active compendiums.
   - Support additional default custom skill (with or without specialisation).
   - Support overriding default test parameters (rollmode, fallback, modifier).
+  - Support for setting group test difficulty in addition to test modifier.
   - Support choice of party (player-assigned characters) or company (non-vehicle player-owned actors) as default player group.
   - Automatically pre-select targets based on tokens selected in scene.
     - If no token is selected, all player group tokens are pre-selected for test. 
     - Otherwise, only selected player and non-player tokens are pre-selected. 
   - Hint prompt to add Specialisation if an undefined grouped skill is selected.
-  - Removed option to set Difficulty for tests (unsupported in system).
   - Removed support for setting SL Bonus or Success Bonus for tests.
 
 ## [Version 0.9.3](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/releases/tag/v0.9.3)  (2022-05-25)